Move ImageLoader to new file
